research/monet-nitrogen/pyproject.toml
[project]
name = "monet-nitrogen"
version = "0.1.0"
description = "Monet Planner + NitroGen Executor: Hierarchical Game AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "numpy>=1.26.0",
    "opencv-python>=4.9.0",
    "pillow>=10.0.0",
    "pydantic>=2.5.0",
    "pyzmq>=25.1.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
windows = [
    "dxcam>=0.0.5",
    "vgamepad>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["schemas", "planner", "executor", "gating", "capture", "orchestrator", "eval"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
